# documentation/swagger.yaml
swagger: "2.0"
info: 
  version: "1.0.0"
  title: GameLabz
  description: |
    ## Required software:

    - Java 1.8+
    - PostgreSQL 9+
    - Apache Maven 3+
    
    ## Build:
    CD to the pom folder
    
    ```
    mvn clean package -DskipTests
    ```
    The jar will be in the /target folder.
    
    This SQL script has to be executed once when the server is deployed for the first time:
    
    ```
    drop database if exists "gameLabz";
    create database "gameLabz";
    ```
    
    The required tables will be created by the application.
    
    ## Usage: 
    ```
    java -jar gameLabz-1.0-SNAPSHOT.jar [parameters]
    Parameters:
    --server.port: the server port, default: 8080
    --database.username: the username for the PostgreSQL database, mandatory
    --database.password: the password for the PostgreSQL database, mandatory
    ```
  contact:
    name: Razman Tamas
host: localhost:8080
basePath: /gameLabz
schemes:
  - http
paths:
...
  /users/{userName}/friendRequest:
    post:
      tags:
        - friendRequest
      summary: A user sends a friend request to another user
      description: After the friend request is being sent the receiver gets a notification on the /friendRequest topic
      operationId: addFriendRequest
      parameters:
        - in: path
          name: userName
          description: Username of the user who wants to send the request
          required: true
          type: string
        - in: header
          name: session_id
          description: It is used to identify the user's session
          required: true
          type: string
        - in: body
          name: friendRequestBody
          description: This contains information about the sender and the receiver
          schema:
            $ref: "#/definitions/FriendRequest"
      responses:
        200:
          description: Friend request has been successfully sent
          schema:
            $ref: "#/definitions/FriendRequestWithId"
        400:
          description: Invalid input
          headers:
            error_message:
              type: string
              description: The error message
        401:
          description: Friend request already exists or user's session is invalid
          headers:
            error_message:
              type: string
              description: The error message
        403:
          description: Couldn't send friend request
          headers:
            error_message:
              type: string
              description: The error message
              
    get:
      tags:
        - friendRequest
      summary: A user lists all friend requests in which he is the sender or the receiver
      description: A friend request is only visible until it is accepted, rejected or cancelled
      operationId: getAllFriendRequests
      parameters:
        - in: path
          name: userName
          description: Username of the user who lists the requests
          required: true
          type: string
        - in: header
          name: session_id
          description: It is used to identify the user's session
          required: true
          type: string
      responses:
        200:
          description: Listing friend requests was succsessful
          schema:
            type: array
            items:
              $ref: "#/definitions/FriendRequestWithId"
        400:
          description: Invalid username
          headers:
            error_message:
              type: string
              description: The error message
        401:
          description: User's session is invalid
          headers:
            error_message:
              type: string
              description: The error message
  /users/{userName}/friendRequest/{friendRequestId}:
    delete:
      tags:
        - friendRequest
      summary: An existing friend request is cancelled by the sender
      description: A friend request can be cancelled before it is accepted or rejected. Cancelling will send a notification to the receiver on the topic /friendRequest 
      operationId: cancelFriendRequest
      parameters:
        - in: path
          name: userName
          description: Username of the user who wants to cancel the request
          required: true
          type: string
        - in: path
          name: friendRequestId
          description: The id of the friend request that is being cancelled
          required: true
          type: integer
          format: int64
        - in: header
          name: session_id
          description: It is used to identify the user's session
          required: true
          type: string
      responses:
        200:
          description: Friend request is successfully cancelled
        400:
          description: Invalid input
          headers:
            error_message:
              type: string
              description: The error message
        401:
          description: User's session is invalid
          headers:
            error_message:
              type: string
              description: The error message
        403: 
          description: This user cannot execute the operation
          headers:
            error_message:
              type: string
              description: The error message
  
    post:
      tags:
        - friendRequest
      summary: A user accepts or rejects a friend request
      description: A user who received a friend request can choose to accept or reject it. The sender will receive a notification on the topic /friendRequest 
      operationId: acceptOrRejectFriendRequest
      parameters:
        - in: query
          name: choice
          description: This parameter is used to decide whether the request is being accepted or rejected 
          required: true
          type: string
          enum:
            - ACCEPTED
            - REJECTED
        - in: path
          name: userName
          description: Username of the user who wants to accept or reject the request
          required: true
          type: string
        - in: path
          name: friendRequestId
          description: The id of the friend request that is being accepted or rejected
          required: true
          type: integer
          format: int64
        - in: header
          name: session_id
          description: It is used to identify the user's session
          required: true
          type: string
      responses:
        200:
          description: The operation was successful
          schema:
            $ref: "#/definitions/FriendRequestWithId"
        400:
          description: Invalid input
          headers:
            error_message:
              type: string
              description: The error message
        401:
          description: User's session is invalid
          headers:
            error_message:
              type: string
              description: The error message
        403: 
          description: This user cannot execute the operation
          headers:
            error_message:
              type: string
              description: The error message
        
        
        
        
#----------------------- GAME RATING ------------------------
...
definitions:
  ResponseUser:
      type: object
      properties:
        userName:
          type: string
        email:
          type: string
        userType:
          type: string
          enum:
            - ADMIN
            - CLIENT
...
  FriendRequestWithId:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fromWho:
          $ref: "#/definitions/ResponseUser"
        toWhom:
          $ref: "#/definitions/ResponseUser"
        status:
          type: string
          enum:
            - CREATED
            - ACCEPTED
            - REJECTED
            - CANCELLED
  FriendRequest:
      type: object
      properties:
        fromWho:
          $ref: "#/definitions/ResponseUser"
        toWhom:
          $ref: "#/definitions/ResponseUser"
        status:
          type: string
          enum:
            - CREATED
            - ACCEPTED
            - REJECTED
            - CANCELLED
...